refactor : (다리 건너기) 다리 연결 메서드 String.join 으로 코드 수정 리팩토링
