+++ src/main/java/bridge/Application.java
@@ -3,7 +3,6 @@ package bridge;
 import bridge.controller.BridgeGameMachine;
 import bridge.view.InputView;
 import bridge.view.OutputView;
-import bridge.view.SystemConsole;
 
 public class Application {
     public static void main(String[] args) {
