+++ src/main/java/bridge/domain/BridgeRandomNumberGenerator.java
@@ -1,6 +1,5 @@
-package bridge;
+package bridge.domain;
 
-import bridge.domain.BridgeNumberGenerator;
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class BridgeRandomNumberGenerator implements BridgeNumberGenerator {
